ui/requirements.txt
bleak==0.22.2
fire==0.6.0
Pillow==10.4.0